Moved util functions from configuration_methods to configuration_utils
