+++ Development/cmake/NmosCppLibraries.cmake
@@ -1007,6 +1007,7 @@ set(NMOS_CPP_NMOS_SOURCES
     nmos/components.cpp
     nmos/configuration_api.cpp
     nmos/configuration_methods.cpp
+	nmos/configuration_utils.cpp
     nmos/connection_activation.cpp
     nmos/connection_api.cpp
     nmos/connection_events_activation.cpp
@@ -1103,6 +1104,7 @@ set(NMOS_CPP_NMOS_HEADERS
     nmos/configuration_api.h
 	nmos/configuration_handlers.h
 	nmos/configuration_methods.h
+	nmos/configuration_utils.h
     nmos/connection_activation.h
     nmos/connection_api.h
     nmos/connection_events_activation.h
