+++ Development/cmake/NmosCppTest.cmake
@@ -43,7 +43,7 @@ set(NMOS_CPP_TEST_NMOS_TEST_SOURCES
     nmos/test/api_utils_test.cpp
     nmos/test/capabilities_test.cpp
     nmos/test/channels_test.cpp
-    nmos/test/configuration_methods_test.cpp
+	nmos/test/configuration_utils_test.cpp
     nmos/test/control_protocol_test.cpp
     nmos/test/did_sdid_test.cpp
     nmos/test/event_type_test.cpp
