+++ Development/nmos/test/configuration_utils_test.cpp
@@ -5,7 +5,7 @@
 #include "nmos/control_protocol_typedefs.h"
 #include "nmos/control_protocol_utils.h"
 #include "nmos/configuration_handlers.h"
-#include "nmos/configuration_methods.h"
+#include "nmos/configuration_utils.h"
 
 #include "bst/test/test.h"
